requirements.txt
asttokens==2.2.1
backcall==0.2.0
bagpy==0.5
bitstring==4.0.2
catkin-pkg==0.5.2
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distro==1.8.0
docutils==0.20.1
exceptiongroup==1.1.1
executing==1.2.0
fonttools==4.39.4
gnupg==2.3.1
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==8.12.2
jedi==0.18.2
Jinja2==3.0.3
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py3rosmsgs==1.18.2
pycryptodomex==3.18.0
Pygments==2.15.1
pyparsing==3.0.9
pyrosbag==0.1.3
pyserial==3.5
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
rospkg==1.5.0
scipy==1.10.1
seaborn==0.12.2
setuptools-scm==7.1.0
six==1.16.0
stack-data==0.6.2
sympy==1.12
tomli==2.0.1
traitlets==5.9.0
typing-extensions==4.6.3
tzdata==2023.3
wcwidth==0.2.6
zipp==3.15.0
